delete `podOwnerHasActiveShares` function and clean up spec files a bit

+++ src/contracts/core/StrategyManager.sol
@@ -202,7 +202,7 @@ contract StrategyManager is
     function withdrawSharesAsTokens(
         address destination,
         IStrategy strategy,
-        uint shares,
+        uint256 shares,
         IERC20 token
     ) external onlyDelegationManager {
         strategy.withdraw(destination, token, shares);
